+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Sermepa - master Oct 2013
 =========================
+- Added a private method to get a camel based method name based on a dash cased name.
 - Added some improvements.
 - Improved exception treatment.
 - Added CHANGELOG.md, support to initialize the instance and updated the demo.